requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
fonttools
gitdb
gitpython
idna
imbalanced-learn
jinja2
joblib
jsonschema
jsonschema-specifications
kiwisolver
markupsafe
matplotlib
narwhals
numpy
packaging
pandas
pillow
protobuf
pyarrow
pydeck
pyparsing
python-dateutil
referencing
requests
rpds-py
scikit-learn==1.6.1
scipy
seaborn
six
sklearn-compat
smmap
streamlit
tenacity
threadpoolctl
toml
tornado
typing-extensions
tzdata
urllib3
watchdog
xgboost
